[CST-7129] Refactoring in newly created ProgrammeCreateGroupConsumer

+++ dspace-api/src/main/java/org/dspace/submit/consumer/ProgrammeCreateGroupConsumer.java
@@ -7,14 +7,24 @@
  */
 package org.dspace.submit.consumer;
 
+import static org.dspace.project.util.ProjectConstants.PROGRAMME;
+import static org.dspace.project.util.ProjectConstants.PROGRAMME_GROUP_TEMPLATE;
+
 import java.io.IOException;
 import java.sql.SQLException;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.Map;
+import java.util.Optional;
+import java.util.Set;
+import java.util.UUID;
 
 import org.dspace.authorize.AuthorizeException;
 import org.dspace.content.DSpaceObject;
 import org.dspace.content.Item;
 import org.dspace.content.factory.ContentServiceFactory;
 import org.dspace.content.service.ItemService;
+import org.dspace.core.Constants;
 import org.dspace.core.Context;
 import org.dspace.eperson.Group;
 import org.dspace.eperson.factory.EPersonServiceFactory;
@@ -30,10 +40,9 @@ import org.dspace.event.Event;
  */
 public class ProgrammeCreateGroupConsumer implements Consumer {
 
-    private final String PROGRAMME_GROUP_NAME = "programme_%s_group";
-
     private ItemService itemService;
     private GroupService groupService;
+    private Map<UUID, Set<Integer>> alreadyProcessed;
 
     @Override
     public void initialize() throws Exception {
@@ -44,18 +53,45 @@ public class ProgrammeCreateGroupConsumer implements Consumer {
     @Override
     public void consume(Context context, Event event) throws Exception {
         DSpaceObject dso = event.getSubject(context);
-        int eventType = event.getEventType();
-        String groupName = String.format(PROGRAMME_GROUP_NAME, event.getSubjectID());
+        if (dso != null && !isItem(dso) || event.getSubjectType() != Constants.ITEM) {
+            return;
+        }
 
-        if (eventType == Event.DELETE) {
+        UUID uuid = event.getSubjectID();
+        int eventType = event.getEventType();
+        String groupName = String.format(PROGRAMME_GROUP_TEMPLATE, uuid.toString());
+
+        boolean processed = hasBeenProcessed(uuid, eventType);
+        if (eventType == Event.DELETE && !processed) {
             deleteGroupByName(context, groupName);
-        } else if (eventType == Event.INSTALL) {
-            if (isEntityTypeEqualTo(dso, "programme")) {
+            putProcessed(uuid, eventType);
+        } else if (eventType == Event.INSTALL && !processed) {
+            Item item = (Item) dso;
+            if (isEntityTypeEqualTo(item, PROGRAMME)) {
                 createProgrammeGroup(context, groupName);
+                putProcessed(uuid, eventType);
             }
         }
 
     }
+
+    @Override
+    public void end(Context ctx) throws Exception {
+        this.alreadyProcessed = null;
+    }
+
+    @Override
+    public void finish(Context ctx) throws Exception {
+
+    }
+
+    private Map<UUID, Set<Integer>> getAlreadyProcessed() {
+        if (this.alreadyProcessed == null) {
+            this.alreadyProcessed = new HashMap<>();
+        }
+        return this.alreadyProcessed;
+    }
+
 
     private void deleteGroupByName(Context context, String groupName)
         throws SQLException, AuthorizeException, IOException {
@@ -65,11 +101,7 @@ public class ProgrammeCreateGroupConsumer implements Consumer {
         }
     }
 
-    private boolean isEntityTypeEqualTo(DSpaceObject dso, String entityType) {
-        if (!isItem(dso)) {
-            return false;
-        }
-        Item item = (Item) dso;
+    private boolean isEntityTypeEqualTo(Item item, String entityType) {
         return itemService.getEntityType(item).equals(entityType);
     }
 
@@ -83,13 +115,22 @@ public class ProgrammeCreateGroupConsumer implements Consumer {
         groupService.update(context, group);
     }
 
-    @Override
-    public void end(Context ctx) throws Exception {
-
+    private void putProcessed(UUID id, Integer event) {
+        Set<Integer> processed = this.getProcessed(id);
+        if (processed == null) {
+            processed = new HashSet<>(2);
+        }
+        processed.add(event);
+        this.getAlreadyProcessed().put(id, processed);
     }
 
-    @Override
-    public void finish(Context ctx) throws Exception {
+    private Set<Integer> getProcessed(UUID id) {
+        return this.getAlreadyProcessed().get(id);
+    }
 
+    private boolean hasBeenProcessed(UUID id, Integer event) {
+        return Optional.ofNullable(this.getAlreadyProcessed().get(id))
+                .map(set -> set.contains(event))
+                .orElse(false);
     }
 }
